start process of rewriting twitter implementation
