+++ package.json
@@ -18,7 +18,7 @@
     "node-schedule": "^2.1.0",
     "source-map-support": "^0.5.21",
     "twitter": "^1.7.1",
-    "twitter.js": "^0.14.0",
+    "twitter-api-v2": "^1.12.3",
     "undici": "^5.3.0",
     "utf-8-validate": "^5.0.9",
     "zlib-sync": "^0.1.7"
